Moved the OptionButton class into the web-components folder.
