+++ src/index.js
@@ -10,6 +10,7 @@ import _Animation from "./_Animation.js";
 
 import Timeline from "./web-components/timeline.js";
 import Toggle from "./web-components/toggle.js";
+import OptionButton from "./web-components/optionButton.js";
 
 class DrawingSelection {
     /** @type {{x: number, y: number}[]} */
@@ -201,35 +202,6 @@ function render(layers = null) {
     }
     rendered.render();
 }
-class OptionButton extends HTMLElement {
-    checked = false;
-    name = "";
-    setChecked(bool) {
-        if (bool === this.checked) {
-            return;
-        }
-        this.onchange();
-        if (bool) {
-            const options = $$(`option-[name="${this.name}"]`);
-            OptionButton.onswitch([...options].indexOf(this), this.name);
-            $$(`option-[name="${this.name}"]`).forEach(el => el.setChecked(false));
-        }
-        this.checked = bool;
-        this.setAttribute("checked", this.checked);
-    }
-    onchange = () => { };
-    static onswitch = () => { };
-    connectedCallback() {
-        this.style.display = 'block';
-        this.name = this.getAttribute("name");
-        this.checked = !!this.getAttribute("checked");
-        this.setAttribute("checked", this.checked);
-        this.onclick = () => {
-            this.setChecked(true);
-        }
-    }
-}
-customElements.define("option-", OptionButton);
 
 let pixelRect;
 let pixelWidth;
